try improve rotation

+++ MalmoExperience/Framework/Block.cs
@@ -1,6 +1,6 @@
 ﻿namespace RunMission.Framework {
     public class Block : BlockOrItem {
-        public Block(double x, double y, double z, 
+        public Block(float x, float y, float z, 
             string name, string variant, 
             Color color, Facing facing)
             : base(x, y, z, name, variant, color, facing) {
