+++ MalmoExperience/Framework/ObservationModels/LineOfSight.cs
@@ -6,11 +6,11 @@ namespace RunMission.Framework.ObservationModels {
         public string HitType { get; set; }
 
         [JsonProperty("x")]
-        public double X { get; set; }
+        public float X { get; set; }
         [JsonProperty("y")]
-        public double Y { get; set; }
+        public float Y { get; set; }
         [JsonProperty("z")]
-        public double Z { get; set; }
+        public float Z { get; set; }
 
         [JsonProperty("type")]
         public string Type { get; set; }
